src/ia/requirements.txt
fastapi
uvicorn
numpy
pydantic>=2
tensorflow==2.20.0
scikit-learn
joblib
python-dotenv
